Moving files and start styling hero
